+++ src/main/java/Application.java
@@ -1,5 +1,3 @@
-import controller.BlackJackGame;
-
 public class Application {
     public static void main(String[] args) {
         BlackJackGame blackJackGame = new BlackJackGame();
