+++ src/main/java/domain/playingcard/Deck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.playingcard;
 
 import java.util.Arrays;
 import java.util.Collections;
