+++ src/main/java/domain/playingcard/PlayingCardShape.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.playingcard;
 
 public enum PlayingCardShape {
     HEART, DIAMOND, CLOVER, SPADE;
