+++ src/main/java/dto/DealerHandStatusDto.java
@@ -1,4 +1,4 @@
-package domain.dto;
+package dto;
 
 import domain.participant.Dealer;
