+++ src/test/java/domain/participant/DealerTest.java
@@ -1,17 +1,16 @@
 package domain.participant;
 
-import domain.Deck;
-import domain.GameResults;
-import domain.PlayingCard;
+import domain.playingcard.Deck;
+import domain.playingcard.PlayingCard;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static domain.PlayingCardShape.DIAMOND;
-import static domain.PlayingCardValue.EIGHT;
-import static domain.PlayingCardValue.NINE;
+import static domain.playingcard.PlayingCardShape.DIAMOND;
+import static domain.playingcard.PlayingCardValue.EIGHT;
+import static domain.playingcard.PlayingCardValue.NINE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class DealerTest {
