+++ src/test/java/domain/playingcard/PlayingCardTest.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.playingcard;
 
+import domain.playingcard.PlayingCard;
+import domain.playingcard.PlayingCardShape;
+import domain.playingcard.PlayingCardValue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,9 +11,9 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-import static domain.PlayingCardShape.HEART;
-import static domain.PlayingCardValue.DEFAULT_ACE;
-import static domain.PlayingCardValue.FIVE;
+import static domain.playingcard.PlayingCardShape.HEART;
+import static domain.playingcard.PlayingCardValue.DEFAULT_ACE;
+import static domain.playingcard.PlayingCardValue.FIVE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class PlayingCardTest {
